+++ astro.config.mjs
@@ -2,7 +2,7 @@ import { defineConfig } from 'astro/config';
 import vercel from '@astrojs/vercel';
 
 export default defineConfig({
-  site: 'https://example.com',
+  site: process.env.SITE_URL,
   output: 'server',
   adapter: vercel(),
   build: {